Use blocks for ManagedObjectUpdatable
