+++ PeakCoreData/Core/Operations/CoreDataSingleImportOperation.swift
@@ -10,8 +10,7 @@ import CoreData
 import PeakOperation
 
 open class CoreDataSingleImportOperation<Intermediate>: CoreDataChangesetOperation, ConsumesResult where
-    Intermediate: ManagedObjectUpdatable & UniqueIdentifiable,
-    Intermediate.ManagedObject: ManagedObjectType & UniqueIdentifiable
+    Intermediate: ManagedObjectUpdatable & UniqueIdentifiable
 {
     public var input: Result<Intermediate, Error> = Result { throw ResultError.noResult }
     
@@ -21,8 +20,8 @@ open class CoreDataSingleImportOperation<Intermediate>: CoreDataChangesetOperati
         do {
             let intermediate = try input.get()
             let managedObject = ManagedObject.fetchOrInsertObject(with: intermediate.uniqueIDValue, in: context, with: cache)
-            intermediate.updateProperties(on: managedObject)
-            intermediate.updateRelationships(on: managedObject, in: context)
+            Intermediate.updateProperties?(intermediate, managedObject)
+            Intermediate.updateRelationships?(intermediate, managedObject, context, cache)
             saveAndFinish()
         } catch {
             output = Result { throw error }
